Issue #11163: Split InputHiddenField1 into focused test inputs
