+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/hiddenfield/InputHiddenFieldEnum.java
@@ -1,0 +1,56 @@
+/*
+HiddenField
+ignoreFormat = (default)null
+ignoreConstructorParameter = (default)false
+ignoreSetter = (default)false
+setterCanReturnItsClass = (default)false
+ignoreAbstractMethods = (default)false
+tokens = VARIABLE_DEF
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.coding.hiddenfield;
+
+enum InputHiddenFieldEnum
+{
+    A(129),
+    B(283),
+    C(1212)
+    {
+        /**
+         * Should not be flagged as violation as we don't check
+         * hidden class level fields
+         */
+        int hidden;
+
+        public void doSomething()
+        {
+            //Should be flagged as hiding enum constant member
+            int hidden = 0; // violation, ''hidden' hides a field'
+        }
+    };
+
+    int hidden;
+    static int hiddenStatic;
+
+    /**
+     * ctor parameter hides member
+     */
+    InputHiddenFieldEnum(int hidden)
+    {
+    }
+
+    public void doSomething()
+    {
+        //Should be flagged as hiding static member
+        int hidden = 0; // violation, ''hidden' hides a field'
+    }
+
+    public static void doSomethingStatic()
+    {
+        //Should be flagged as hiding static member
+        int hiddenStatic = 0; // violation, ''hiddenStatic' hides a field'
+    }
+}
+
