+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/hiddenfield/InputHiddenFieldPropertySetter.java
@@ -1,0 +1,94 @@
+/*
+HiddenField
+ignoreFormat = (default)null
+ignoreConstructorParameter = (default)false
+ignoreSetter = (default)false
+setterCanReturnItsClass = (default)false
+ignoreAbstractMethods = (default)false
+tokens = VARIABLE_DEF
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.coding.hiddenfield;
+
+/** tests ignoring the parameter of a property setter method */
+class InputHiddenFieldPropertySetter
+{
+    private int prop;
+
+    /** setter */
+    public void setProp(int prop)
+    {
+        this.prop = prop;
+    }
+
+    /** violation - incorrect method name */
+    public void setprop(int prop)
+    {
+        this.prop = prop;
+    }
+
+    /** violation - more than one parameter */
+    public void setProp(int prop, int extra)
+    {
+        this.prop = prop;
+    }
+}
+
+/** tests a non-void method */
+class PropertySetter21
+{
+    private int prop;
+
+    /** violation - not a void method */
+    public int setProp(int prop)
+    {
+        this.prop = prop;
+        return 0;
+    }
+}
+
+/** tests chain-setter */
+class PropertySetter31
+{
+    private int prop;
+
+    /**
+     * if setterCanReturnItsClass == false then
+     *     violation - not a void method
+     *
+     * if setterCanReturnItsClass == true then
+     *     success as it is then considered to be a setter
+     */
+    public PropertySetter31 setProp(int prop)
+    {
+        this.prop = prop;
+        return this;
+    }
+}
+
+/** tests setters (both regular and the chain one) on the enum */
+enum PropertySetter41 {
+    INSTANCE;
+
+    private int prop;
+    private int prop2;
+
+    public void setProp(int prop) {
+        this.prop = prop;
+    }
+
+    /**
+     * if setterCanReturnItsClass == false then
+     *     violation - not a void method
+     *
+     * if setterCanReturnItsClass == true then
+     *     success as it is then considered to be a setter
+     */
+    public PropertySetter41 setProp2(int prop2)
+    {
+        this.prop2 = prop2;
+        return this;
+    }
+}
